+++ lib/deltaq/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Revision history for `deltaq`
 
-## 1.0.2.0 — to be released
+## 1.1.0.0 — to be released
+
+### Added
 
 * Add `DeltaQ.Diagram` for rendering outcome expressions as diagrams.
 * Add `DeltaQ.Expr` with a data type `O` for outcome expressions.
+
+### Changed
+
+* Add the new class `ProbabilisticOutcome` and remove its methods
+  `Probability`, `choice` and `choices` from the `DeltaQ` class.
 
 ## 1.0.1.0 — 2025-09-26